refactor : 업로드 컴포넌트 파일 위치 변경
